# api/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentação da API de delivery Cardapiogo.",
        "title": "Cardapiogo API",
        "contact": {
            "name": "Gustavo Ferreira de Jesus",
            "email": "gustavo081900@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/v1/store": {
            "get": {
                "description": "Returns a list of stores based on various filter criteria including location, rating, and type.\nSupports geographic search with radius filtering when latitude and longitude are provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Search and filter stores",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum distance range in kilometers for geographical search",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum rating score (1-5)",
                        "name": "score",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Store name for partial matching",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by city name",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Latitude for geographical search (requires longitude)",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Longitude for geographical search (requires latitude)",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Store type category (e.g., restaurant, cafe, bar)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of stores matching the filter criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.GetStoreByIdOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter parameters",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "404": {
                        "description": "No stores found",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Allows store owners to modify their store's details and settings.\nAll provided fields will be updated while maintaining existing data for omitted fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Update store information",
                "parameters": [
                    {
                        "description": "Updated store information including modifiable fields",
                        "name": "Params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.UpdateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store successfully updated"
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Bearer token missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not the store owner",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Store not found",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict with existing data",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Allows authenticated users to register a new store in the system.\nThe store owner must provide essential information including name, location, and contact details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Create a new store",
                "parameters": [
                    {
                        "description": "Store creation parameters including name, address, contact info, and type",
                        "name": "Params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.CreateParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns the created store ID",
                        "schema": {
                            "$ref": "#/definitions/store.CreatedStore"
                        }
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Bearer token missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User lacks permission",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict - Store already exists",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
...
        "/v1/store/{id}": {
            "get": {
                "description": "Returns detailed information about a store by its ID.\nThis endpoint is publicly accessible and provides store details including business hours,\ncontact information, location, and current status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Retrieve a specific store",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store ID - Unique identifier of the store",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store details including name, description, business hours, and contact info",
                        "schema": {
                            "$ref": "#/definitions/store.GetStoreByIdOutput"
                        }
                    },
                    "404": {
                        "description": "Store not found",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
        "/v1/store/{id}/header-image": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Allows store owners to upload or update their store's header/banner image.\nSupports JPEG and PNG formats. Image will be processed and optimized for display.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Upload store header image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Header image file (JPEG/PNG only, max size: 5MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns URL of the uploaded image",
                        "schema": {
                            "$ref": "#/definitions/controller.setFileOutput"
                        }
                    },
                    "400": {
                        "description": "Invalid file format or size",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Bearer token missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not the store owner",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "413": {
                        "description": "File size exceeds limit",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.setFileOutput": {
            "type": "object",
            "properties": {
                "objectURL": {
                    "type": "string",
                    "example": "URL of object on storage"
                }
            }
        },
...
        "store.CreatedStore": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "store.GetStoreByIdOutput": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/store.AddressOutput"
                },
                "businessHours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.BusinessHoursParams"
                    }
                },
                "headerImage": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paymentMethod": {
                    "type": "array",
                    "items": {
                        "enum": [
                            "credit",
                            " debit",
                            " pix",
                            " cash"
                        ],
                        "$ref": "#/definitions/store.PaymentMethod"
                    }
                },
                "phone": {
                    "type": "string"
                },
                "profileImage": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "type": {
                    "enum": [
                        "restaurant",
                        " pharmacy",
                        " tobbaco",
                        " market",
                        " convenience",
                        " pub"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/store.ShopType"
                        }
                    ]
                }
            }
        },
        "store.PaymentMethod": {
            "type": "string",
            "enum": [
                "credit",
                "debit",
                "pix",
                "cash"
            ],
            "x-enum-varnames": [
                "Credit",
                "Debit",
                "Pix",
                "Cash"
            ]
        },
...
        "xerrors.CustomError": {
            "type": "object",
            "properties": {
                "debug": {},
                "details": {},
                "error": {
                    "type": "string"
                },
                "traceID": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}